+++ myqueue/test/test_workflow.py
@@ -2,23 +2,10 @@ import os
 from pathlib import Path
 from myqueue.task import task
 from myqueue.test.flow1 import workflow
-# from myqueue.workflow import workflow_from_function
 
 
 def create_tasks():
     return [task('somepackage.somemodule')]
-
-
-def xxxtest_basic_workflow():
-    """
-    tasks = workflow_from_function(
-        workflow_function=create_tasks,
-        folders=[Path('.')])
-
-    assert tasks[0].todict() == task('somepackage.somemodule',
-                                     workflow=True).todict()
-    """
-    pass
 
 
 def test_flow1(mq):